requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
Deprecated==1.2.10
Django==3.1.4
idna==2.10
Markdown==3.3.3
markdown2==2.3.10
numpy==1.19.4
Pillow==8.0.1
PyGithub==1.54
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
requests==2.24.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.25.11
wget==3.2
wrapt==1.12.1
